Refactor : thank you pages
