+++ vue-frontend/src/components/contact/thank-you/BenefitSection.vue
