+++ vue-frontend/src/components/jobs/thank-you/ContributionSection.vue
