+++ vue-frontend/src/components/jobs/thank-you/LatestBlog.vue
